convert Manhwa18Net to factory
