+++ multisrc/overrides/fmreader/manhwa18net/src/Manhwa18NetFactory.kt
@@ -5,6 +5,13 @@ import eu.kanade.tachiyomi.network.GET
 import eu.kanade.tachiyomi.source.model.FilterList
 import okhttp3.Request
 
+class Manhwa18NetFactory : SourceFactory {
+    override fun createSources(): List<Source> = listOf(
+            Manhwa18Net(),
+            Manhwa18NetRaw(),
+    )
+}
+
 class Manhwa18Net : FMReader("Manhwa18.net", "https://manhwa18.net", "en") {
     override fun popularMangaRequest(page: Int): Request =
             GET("$baseUrl/$requestPath?listType=pagination&page=$page&sort=views&sort_type=DESC&ungenre=raw", headers)
@@ -18,4 +25,14 @@ class Manhwa18Net : FMReader("Manhwa18.net", "https://manhwa18.net", "en") {
     }
 
     override fun getGenreList() = getAdultGenreList()
+}
+
+class Manhwa18NetRaw : FMReader("Manhwa18.net", "https://manhwa18.net", "ko") {
+    override val requestPath = "manga-list-genre-raw.html"
+    override fun searchMangaRequest(page: Int, query: String, filters: FilterList): Request {
+        val onlyRawsUrl = super.searchMangaRequest(page, query, filters).url().newBuilder().addQueryParameter("genre", "raw").toString()
+        return GET(onlyRawsUrl, headers)
+    }
+
+    override fun getFilterList() = FilterList(super.getFilterList().filterNot { it == GenreList(getGenreList()) })
 }
